signals

+++ stats_functions.c
@@ -360,7 +360,7 @@ void getMemoryUsage(int write_pipe)
     free(buf);
 }
 
-void getMemoryUsageGraphic(int current_usage, int previous_usage)
+void getMemoryUsageGraphic(float current_usage, float previous_usage)
 {
     // This function takes the current memory usage (float current_usage) and previous usage (float previous_usage) and
     // formats a graphic for the current memory usage. The function will return the properly formatted string that includes
@@ -406,7 +406,7 @@ void getMemoryUsageGraphic(int current_usage, int previous_usage)
     }
 
     // add the current usage
-    sprintf(buf + strlen(buf), " %.2f (%.2f)", (double)difference, current_usage);
+    sprintf(buf + strlen(buf), " %.2f (%.2f)", difference, current_usage);
 
     printf("%s \n", buf);
 }
